requirements.txt
fastapi
uvicorn
requests
cryptography
python-dotenv

